+++ core/src/main/java/com/minekube/connect/tunnel/p2p/Libp2pTunnelTransport.java
@@ -26,155 +26,34 @@
 package com.minekube.connect.tunnel.p2p;
 
 import com.google.inject.Inject;
-import com.minekube.connect.api.logger.ConnectLogger;
-import com.minekube.connect.tunnel.P2PTunnelHeader;
 import com.minekube.connect.tunnel.TunnelClientTransport;
 import com.minekube.connect.tunnel.TunnelConn;
-import io.libp2p.core.Connection;
-import io.libp2p.core.Host;
-import io.libp2p.core.PeerId;
-import io.libp2p.core.crypto.PrivKey;
-import io.libp2p.core.Stream;
-import io.libp2p.core.StreamPromise;
-import io.libp2p.core.crypto.KeyType;
-import io.libp2p.core.dsl.HostBuilder;
-import io.libp2p.core.mux.StreamMuxerProtocol;
-import io.libp2p.core.multiformats.Multiaddr;
-import io.libp2p.core.multistream.StrictProtocolBinding;
-import io.libp2p.protocol.circuit.CircuitHopProtocol;
-import io.libp2p.protocol.circuit.CircuitStopProtocol;
-import io.libp2p.protocol.circuit.RelayTransport;
-import io.libp2p.protocol.ProtocolHandler;
-import io.libp2p.security.noise.NoiseXXSecureChannel;
-import io.libp2p.transport.ConnectionUpgrader;
-import io.libp2p.transport.quic.QuicConfig;
-import io.libp2p.transport.quic.QuicTransport;
-import io.libp2p.transport.tcp.TcpTransport;
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.ByteBufUtil;
-import io.netty.buffer.Unpooled;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.SimpleChannelInboundHandler;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeoutException;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 import minekube.connect.v1alpha1.WatchServiceOuterClass.TunnelTransport.Type;
 
 public final class Libp2pTunnelTransport implements TunnelClientTransport {
-    public static final String PROTOCOL_ID = "/minekube/connect/tunnel/1.0.0";
-
-    private static final long START_TIMEOUT_SECONDS = 10;
-    private static final long CONNECT_TIMEOUT_SECONDS = 15;
-    private static final long STREAM_TIMEOUT_SECONDS = 5;
-
-    private final Host host;
-    private final ConnectLogger logger;
-    private final ConcurrentMap<PeerId, Connection> warmConnections = new ConcurrentHashMap<>();
-    private boolean started;
-
-    public Libp2pTunnelTransport() {
-        this(createHost(), null);
-    }
+    private final TunnelClientTransport delegate;
+    private final RuntimeException unavailable;
 
     @Inject
-    public Libp2pTunnelTransport(ConnectLogger logger) {
-        this(createHost(), logger);
-    }
-
-    Libp2pTunnelTransport(Host host) {
-        this(host, null);
-    }
-
-    private Libp2pTunnelTransport(Host host, ConnectLogger logger) {
-        this.host = Objects.requireNonNull(host, "host");
-        this.logger = logger;
-        this.host.addProtocolHandler(new TunnelProtocolBinding());
-    }
-
-    static Host createHost() {
-        return createHost(null, new String[0]);
-    }
-
-    static Host createHost(PrivKey privateKey) {
-        return createHost(privateKey, new String[0]);
-    }
-
-    static Host createHost(PrivKey privateKey, String... listenAddrs) {
-        return createHost(privateKey, listenAddrs, Collections.emptyList());
-    }
-
-    static Host createHost(PrivKey privateKey, String[] listenAddrs, List<String> relayAddrs) {
-        return createHost(privateKey, listenAddrs, relayAddrs, false);
-    }
-
-    static Host createRelayServiceHost(PrivKey privateKey, String... listenAddrs) {
-        return createHost(privateKey, listenAddrs, Collections.emptyList(), true);
-    }
-
-    private static Host createHost(
-            PrivKey privateKey,
-            String[] listenAddrs,
-            List<String> relayAddrs,
-            boolean relayService) {
-        RelayBindings relay = relayBindings(privateKey, relayAddrs, relayService);
-        HostBuilder builder = new HostBuilder(HostBuilder.DefaultMode.None)
-                .secureChannel(NoiseXXSecureChannel::new)
-                .muxer(StreamMuxerProtocol::getYamux)
-                .secureTransport((priv, protocols) ->
-                        QuicTransport.Ed25519(priv, protocols, new QuicConfig()));
-        if (relay == null) {
-            builder.transport(TcpTransport::new);
-        } else {
-            builder.transport(TcpTransport::new, relay::transport);
-            builder.protocol(relay.hopBinding, relay.stopBinding);
+    public Libp2pTunnelTransport() {
+        TunnelClientTransport created = null;
+        RuntimeException failure = null;
+        try {
+            Class<?> runtimeClass = Class.forName(
+                    "com.minekube.connect.tunnel.p2p.impl.Libp2pTunnelTransportRuntime",
+                    true,
+                    Libp2pRuntimeLoader.classLoader());
+            created = (TunnelClientTransport) runtimeClass.getDeclaredConstructor().newInstance();
+        } catch (Exception | LinkageError e) {
+            failure = new IllegalStateException("Connect libp2p transport runtime unavailable", e);
         }
-        if (listenAddrs != null && listenAddrs.length > 0) {
-            builder.listen(listenAddrs);
-        }
-        Host built = privateKey == null
-                ? builder
-                        .keyType(KeyType.ED25519)
-                        .build()
-                : builder
-                        .builderModifier(builderJ -> builderJ.getIdentity().setFactory(() -> privateKey))
-                        .build();
-        if (relay != null) {
-            relay.attachHost(built);
-        }
-        return built;
+        this.delegate = created;
+        this.unavailable = failure;
     }
 
-    private static RelayBindings relayBindings(PrivKey privateKey, List<String> relayAddrs, boolean relayService) {
-        if (!relayService && (relayAddrs == null || relayAddrs.isEmpty())) {
-            return null;
-        }
-        CircuitStopProtocol.Binding stopBinding = new CircuitStopProtocol.Binding(new CircuitStopProtocol());
-        CircuitHopProtocol.RelayManager relayManager = relayService
-                ? CircuitHopProtocol.RelayManager.limitTo(
-                        privateKey,
-                        PeerId.fromPubKey(privateKey.publicKey()),
-                        1024)
-                : new DenyRelayManager();
-        CircuitHopProtocol.Binding hopBinding = new CircuitHopProtocol.Binding(relayManager, stopBinding);
-        List<RelayTransport.CandidateRelay> candidates = relayAddrs == null
-                ? Collections.emptyList()
-                : relayAddrs.stream()
-                        .map(Multiaddr::fromString)
-                        .map(addr -> new RelayTransport.CandidateRelay(
-                                requirePeerId(addr, addr.toString()),
-                                Collections.singletonList(addr)))
-                        .collect(Collectors.toList());
-        return new RelayBindings(hopBinding, stopBinding, candidates);
+    Libp2pTunnelTransport(TunnelClientTransport delegate) {
+        this.delegate = delegate;
+        this.unavailable = null;
     }
 
     @Override
@@ -184,256 +63,23 @@ public final class Libp2pTunnelTransport implements TunnelClientTransport {
 
     @Override
     public void prepare(String address) {
-        if (address == null || address.isEmpty()) {
-            return;
+        if (delegate != null) {
+            delegate.prepare(address);
         }
-
-        Multiaddr multiaddr = Multiaddr.fromString(address);
-        PeerId peerId = requirePeerId(multiaddr, address);
-        warmConnections.compute(peerId, (ignored, existing) -> {
-            if (isHealthy(existing)) {
-                return existing;
-            }
-            return connect(peerId, multiaddr);
-        });
-    }
-
-    boolean hasWarmConnection(String address) {
-        Multiaddr multiaddr = Multiaddr.fromString(address);
-        PeerId peerId = requirePeerId(multiaddr, address);
-        return isHealthy(warmConnections.get(peerId));
     }
 
     @Override
     public TunnelConn tunnel(String address, String sessionId, TunnelConn.Handler handler) {
-        Objects.requireNonNull(handler, "handler");
-        byte[] header = P2PTunnelHeader.encode(sessionId);
-        Multiaddr multiaddr = Multiaddr.fromString(address);
-        PeerId peerId = requirePeerId(multiaddr, address);
-        Connection connection = warmConnection(peerId, multiaddr);
-        Stream stream = openStream(connection);
-        try {
-            stream.pushHandler(new InboundBytesHandler(handler, logger));
-            stream.writeAndFlush(Unpooled.wrappedBuffer(header));
-            return new Libp2pTunnelConn(stream, logger);
-        } catch (RuntimeException e) {
-            stream.close();
-            throw e;
+        if (delegate == null) {
+            throw unavailable;
         }
+        return delegate.tunnel(address, sessionId, handler);
     }
 
     @Override
     public void close() {
-        warmConnections.clear();
-        Host host;
-        synchronized (this) {
-            if (!started) {
-                return;
-            }
-            host = this.host;
-            started = false;
-        }
-        await(host.stop(), START_TIMEOUT_SECONDS, "stop libp2p host");
-    }
-
-    private Connection warmConnection(PeerId peerId, Multiaddr multiaddr) {
-        prepare(multiaddr.toString());
-        Connection connection = warmConnections.get(peerId);
-        if (!isHealthy(connection)) {
-            throw new IllegalStateException("libp2p connection is not warm for " + multiaddr);
-        }
-        return connection;
-    }
-
-    private Connection connect(PeerId peerId, Multiaddr multiaddr) {
-        return await(startedHost().getNetwork().connect(peerId, multiaddr),
-                CONNECT_TIMEOUT_SECONDS, "connect libp2p peer " + peerId);
-    }
-
-    private Stream openStream(Connection connection) {
-        StreamPromise<Object> promise = startedHost().newStream(Arrays.asList(PROTOCOL_ID), connection);
-        Stream stream = await(promise.getStream(), STREAM_TIMEOUT_SECONDS, "open libp2p tunnel stream");
-        await(stream.getProtocol(), STREAM_TIMEOUT_SECONDS, "negotiate libp2p tunnel protocol");
-        return stream;
-    }
-
-    private synchronized Host startedHost() {
-        if (!started) {
-            await(host.start(), START_TIMEOUT_SECONDS, "start libp2p host");
-            started = true;
-        }
-        return host;
-    }
-
-    private static PeerId requirePeerId(Multiaddr multiaddr, String address) {
-        PeerId peerId = multiaddr.getPeerId();
-        if (peerId == null) {
-            throw new IllegalArgumentException("p2p address must include /p2p/<peer-id>: " + address);
-        }
-        return peerId;
-    }
-
-    private static boolean isHealthy(Connection connection) {
-        return connection != null && !connection.closeFuture().isDone();
-    }
-
-    private static final class RelayBindings {
-        private final CircuitHopProtocol.Binding hopBinding;
-        private final CircuitStopProtocol.Binding stopBinding;
-        private final List<RelayTransport.CandidateRelay> candidates;
-        private final ScheduledExecutorService executor = Executors.newSingleThreadScheduledExecutor(runnable -> {
-            Thread thread = new Thread(runnable, "connect-libp2p-relay");
-            thread.setDaemon(true);
-            return thread;
-        });
-        private RelayTransport transport;
-
-        private RelayBindings(
-                CircuitHopProtocol.Binding hopBinding,
-                CircuitStopProtocol.Binding stopBinding,
-                List<RelayTransport.CandidateRelay> candidates) {
-            this.hopBinding = hopBinding;
-            this.stopBinding = stopBinding;
-            this.candidates = candidates;
-        }
-
-        private RelayTransport transport(ConnectionUpgrader upgrader) {
-            RelayTransport transport = new RelayTransport(
-                    hopBinding,
-                    stopBinding,
-                    upgrader,
-                    ignored -> candidates,
-                    executor);
-            transport.setRelayCount(candidates.size());
-            this.transport = transport;
-            return transport;
-        }
-
-        private void attachHost(Host host) {
-            if (transport != null) {
-                transport.setHost(host);
-            } else {
-                hopBinding.setHost(host);
-            }
-        }
-    }
-
-    private static final class DenyRelayManager implements CircuitHopProtocol.RelayManager {
-        @Override
-        public boolean hasReservation(PeerId peerId) {
-            return false;
-        }
-
-        @Override
-        public Optional<CircuitHopProtocol.Reservation> createReservation(PeerId peerId, Multiaddr multiaddr) {
-            return Optional.empty();
-        }
-
-        @Override
-        public Optional<CircuitHopProtocol.Reservation> allowConnection(PeerId source, PeerId destination) {
-            return Optional.empty();
-        }
-    }
-
-    private static <T> T await(CompletableFuture<T> future, long timeoutSeconds, String action) {
-        try {
-            return future.get(timeoutSeconds, TimeUnit.SECONDS);
-        } catch (InterruptedException e) {
-            future.cancel(true);
-            Thread.currentThread().interrupt();
-            throw new IllegalStateException("failed to " + action, e);
-        } catch (TimeoutException e) {
-            future.cancel(true);
-            throw new IllegalStateException("failed to " + action, e);
-        } catch (Exception e) {
-            throw new IllegalStateException("failed to " + action, e);
-        }
-    }
-
-    private static final class Libp2pTunnelConn extends TunnelConn {
-        private final Stream stream;
-        private final ConnectLogger logger;
-
-        private Libp2pTunnelConn(Stream stream, ConnectLogger logger) {
-            this.stream = stream;
-            this.logger = logger;
-        }
-
-        @Override
-        public void write(byte[] data) {
-            stream.writeAndFlush(Unpooled.wrappedBuffer(data));
-        }
-
-        @Override
-        public void close(Throwable t) {
-            if (logger != null) {
-                if (t == null) {
-                    logger.info("Connect libp2p tunnel close requested by local backend channel");
-                } else {
-                    logger.warn("Connect libp2p tunnel close requested with cause: " + t);
-                }
-            }
-            stream.close();
-        }
-
-        @Override
-        public boolean opened() {
-            return true;
-        }
-    }
-
-    private static final class InboundBytesHandler extends SimpleChannelInboundHandler<ByteBuf> {
-        private final TunnelConn.Handler handler;
-        private final ConnectLogger logger;
-
-        private InboundBytesHandler(TunnelConn.Handler handler, ConnectLogger logger) {
-            this.handler = handler;
-            this.logger = logger;
-        }
-
-        @Override
-        protected void channelRead0(ChannelHandlerContext ctx, ByteBuf msg) {
-            handler.onReceive(ByteBufUtil.getBytes(msg, msg.readerIndex(), msg.readableBytes(), false));
-        }
-
-        @Override
-        public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) {
-            if (logger != null) {
-                logger.warn("Connect libp2p tunnel inbound stream error: " + cause);
-            }
-            handler.onError(cause);
-            ctx.close();
-        }
-
-        @Override
-        public void channelInactive(ChannelHandlerContext ctx) throws Exception {
-            if (logger != null) {
-                logger.info("Connect libp2p tunnel inbound stream closed; closing local backend channel");
-            }
-            handler.onClose();
-            super.channelInactive(ctx);
-        }
-    }
-
-    private static final class TunnelProtocolBinding extends StrictProtocolBinding<Void> {
-        private TunnelProtocolBinding() {
-            super(PROTOCOL_ID, new TunnelProtocolHandler());
-        }
-    }
-
-    private static final class TunnelProtocolHandler extends ProtocolHandler<Void> {
-        private TunnelProtocolHandler() {
-            super(Long.MAX_VALUE, Long.MAX_VALUE);
-        }
-
-        @Override
-        protected CompletableFuture<Void> onStartInitiator(Stream stream) {
-            return CompletableFuture.completedFuture(null);
-        }
-
-        @Override
-        protected CompletableFuture<Void> onStartResponder(Stream stream) {
-            return CompletableFuture.completedFuture(null);
+        if (delegate != null) {
+            delegate.close();
         }
     }
 }
