+++ core/src/main/java/com/minekube/connect/tunnel/p2p/PeerRegistrationClient.java
@@ -64,7 +64,7 @@ final class PeerRegistrationClient {
             List<String> observedAddrs,
             long sequence,
             long nowUnixMs) {
-        return installResolved(stream, observedAddrs, sequence, nowUnixMs);
+        return install(stream, () -> observedAddrs, sequence, nowUnixMs);
     }
 
     CompletableFuture<PeerRegisterResult> install(
@@ -72,16 +72,8 @@ final class PeerRegistrationClient {
             Supplier<List<String>> observedAddrsSupplier,
             long sequence,
             long nowUnixMs) {
-        List<String> observedAddrs = observedAddrsSupplier.get();
-        return installResolved(stream, observedAddrs, sequence, nowUnixMs);
-    }
-
-    private CompletableFuture<PeerRegisterResult> installResolved(
-            Stream stream,
-            List<String> observedAddrs,
-            long sequence,
-            long nowUnixMs) {
         this.stream = stream;
+        List<String> observedAddrs = observedAddrsSupplier.get();
         CompletableFuture<PeerRegisterResult> result = new CompletableFuture<>();
         P2PFrameDecoder<PeerRegisterChallenge> challengeDecoder = new P2PFrameDecoder<>(
                 PeerRegisterChallenge.parser(),
@@ -90,6 +82,7 @@ final class PeerRegistrationClient {
         stream.pushHandler(new ChallengeHandler(
                 stream,
                 challengeDecoder,
+                observedAddrsSupplier,
                 observedAddrs,
                 sequence,
                 nowUnixMs,
@@ -115,14 +108,14 @@ final class PeerRegistrationClient {
             ChannelHandlerContext ctx,
             Stream stream,
             PeerRegisterChallenge challenge,
-            List<String> observedAddrs,
+            Supplier<List<String>> observedAddrsSupplier,
             long sequence,
             CompletableFuture<PeerRegisterResult> result) {
         P2PFrameDecoder<PeerRegisterResult> resultDecoder = new P2PFrameDecoder<>(
                 PeerRegisterResult.parser(),
                 P2PFrameCodec.MAX_CONTROL_FRAME_SIZE);
         ctx.pipeline().addLast(resultDecoder);
-        ctx.pipeline().addLast(new ResultHandler(stream, challenge, observedAddrs, sequence, result));
+        ctx.pipeline().addLast(new ResultHandler(stream, challenge, observedAddrsSupplier, sequence, result));
     }
 
     private static void writeFrame(Stream stream, MessageLite message) {
@@ -138,6 +131,7 @@ final class PeerRegistrationClient {
     private final class ChallengeHandler extends SimpleChannelInboundHandler<PeerRegisterChallenge> {
         private final Stream stream;
         private final ChannelHandler decoder;
+        private final Supplier<List<String>> observedAddrsSupplier;
         private final List<String> observedAddrs;
         private final long sequence;
         private final long nowUnixMs;
@@ -146,12 +140,14 @@ final class PeerRegistrationClient {
         private ChallengeHandler(
                 Stream stream,
                 ChannelHandler decoder,
+                Supplier<List<String>> observedAddrsSupplier,
                 List<String> observedAddrs,
                 long sequence,
                 long nowUnixMs,
                 CompletableFuture<PeerRegisterResult> result) {
             this.stream = stream;
             this.decoder = decoder;
+            this.observedAddrsSupplier = observedAddrsSupplier;
             this.observedAddrs = observedAddrs;
             this.sequence = sequence;
             this.nowUnixMs = nowUnixMs;
@@ -163,7 +159,7 @@ final class PeerRegistrationClient {
             ctx.pipeline().remove(this);
             ctx.pipeline().remove(decoder);
             writeFrame(stream, handshake.commit(challenge, observedAddrs, sequence, nowUnixMs));
-            installResultHandler(ctx, stream, challenge, observedAddrs, sequence, result);
+            installResultHandler(ctx, stream, challenge, observedAddrsSupplier, sequence, result);
         }
 
         @Override
@@ -188,7 +184,7 @@ final class PeerRegistrationClient {
     private final class ResultHandler extends SimpleChannelInboundHandler<PeerRegisterResult> {
         private final Stream stream;
         private final PeerRegisterChallenge challenge;
-        private final List<String> observedAddrs;
+        private final Supplier<List<String>> observedAddrsSupplier;
         private final AtomicLong sequence;
         private final CompletableFuture<PeerRegisterResult> result;
         private volatile ScheduledFuture<?> ackTimeout;
@@ -196,12 +192,12 @@ final class PeerRegistrationClient {
         private ResultHandler(
                 Stream stream,
                 PeerRegisterChallenge challenge,
-                List<String> observedAddrs,
+                Supplier<List<String>> observedAddrsSupplier,
                 long sequence,
                 CompletableFuture<PeerRegisterResult> result) {
             this.stream = stream;
             this.challenge = challenge;
-            this.observedAddrs = observedAddrs;
+            this.observedAddrsSupplier = observedAddrsSupplier;
             this.sequence = new AtomicLong(sequence);
             this.result = result;
         }
@@ -219,7 +215,7 @@ final class PeerRegistrationClient {
                     try {
                         writeFrame(stream, handshake.commit(
                                 challenge,
-                                observedAddrs,
+                                observedAddrsSupplier.get(),
                                 sequence.incrementAndGet(),
                                 System.currentTimeMillis()));
                         scheduleAckTimeout();
